nav bar styles and form editing
added a content component that is used to desplay the content of a page in the main layout.
main layout is used to define the layout of the site mainly the nav bar being displayed next to the page content.
also added some styles to the nav bar, including functionality for highlighting current page and hovered over link